requirements.txt
absl-py==0.11.0
dm-sonnet==2.0.0
dm-tree==0.1.5
git+git://github.com/google/gin-config.git@master#egg=gin-config
gym==0.17.3
h5py==2.10.0
mpyq==0.2.5
numpy==1.18.5
pypeln==0.4.7
git+git://github.com/metataro/pysc2.git@master#egg=PySC2
PyYAML==5.3.1
s2clientprotocol==5.0.2.81102.0
s2protocol==5.0.3.81433.0
sc2reader==1.6.0
scipy==1.5.3
tensorboard==2.3.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-probability==0.11.1
tqdm==4.51.0
wandb==0.10.12
